Switching out how headers are handled, and updating the permissions test.
